refactor(team): extract entity types into broker_types.go
Per staff-reviewer's "if you wanted a fourth split, broker_types.go
would do it. I'd actually do this — having every entity type defined
in one alphabetised file is much easier to skim than them being
cross-referenced from a 1500-line file."

Moved 21 wire-shape entity types (the JSON-tagged structs the broker
persists and serves over HTTP) into broker_types.go. The header
comment maps each type to its coupling: brokerState composes
everything; usageTotals/teamUsageState track aggregates;
officeMember.Provider feeds the launcher dispatch.

Types moved:
  messageReaction, channelMessage, agentIssueRecord, messageUsage,
  interviewOption, interviewAnswer, humanInterview (with
  TitleOrDefault method — the only formatter belonging to a type
  rather than to an entity file), teamTask, channelSurface,
  teamChannel, officeMember, officeActionLog,
  agentActivitySnapshot, officeSignalRecord, officeDecisionRecord,
  watchdogAlert, schedulerJob, teamSkill, brokerState, usageTotals,
  teamUsageState

What stays in broker.go: ipRateLimitBucket (tightly bound to the
runtime rate-limiter), Broker struct (bound to constructor + every
method receiver), constants, and lifecycle.

  - internal/team/broker.go: -319 LOC (1207 -> 888)
  - internal/team/broker_types.go: +345 LOC (new)

Cumulative broker.go reduction:
  3404 -> 888
  Net: -2516 LOC; broker.go shrunk by 74%.

The remaining 888-LOC broker.go is the genuine "core": Broker struct +
constructors (NewBroker / NewBrokerAt) + lifecycle (Start /
StartOnPort / Stop / Reset) + ipRateLimitBucket + access primitives
(findChannelLocked / findMemberLocked /
ensureDMConversationLocked / rebuildMemberIndexLocked /
appendActionLocked) + session-mode setters + openclaw bridge attach
+ ExternalQueue / EnsureBridgedMember / EnsureDirectChannel /
PostInboundSurfaceMessage (the side-effecting "ensure" methods that
the staff-reviewer correctly noted are queries in name only).

Tests pass.
